requirements.txt
absolufy-imports==0.3.0
appdirs==1.4.4
beautifulsoup4==4.9.3
black==21.6b0
cchardet==2.1.7
Cerberus==1.3.4
certifi==2021.5.30
cffi==1.14.5
cfgv==3.3.0
chardet==4.0.0
click==8.0.1
Deprecated==1.2.12
distlib==0.3.2
filelock==3.0.12
Flask==2.0.1
Flask-Assets==2.0
Flask-JSON==0.3.4
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
identify==2.2.10
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
nodeenv==1.6.0
numpy==1.21.0
pathspec==0.8.1
pre-commit==2.13.0
psycopg2==2.9.1
pycparser==2.20
pycraigslist==0.5.3
PyGithub==1.55
pyinstrument==3.4.2
pyinstrument-cext==0.2.4
PyJWT==2.1.0
PyMySQL==1.0.2
PyNaCl==1.4.0
PyYAML==5.4.1
regex==2021.7.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.20
tenacity==8.0.0
toml==0.10.2
urllib3==1.26.6
virtualenv==20.4.7
webassets==2.0
Werkzeug==2.0.1
wrapt==1.12.1
